Remove version requirement for Firefox/Chrome with the goal of always tracking the latest version.
It should be noted the Chrome repository does not keep older versions available once a new one is released, so this more of a requirement to make things work in case the box needs to be updated. For Firefox the situation isn't as such because older packages are still made available by Fedora.